requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
django-rest==0.8.6
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
